1차 합산본

+++ src/main/java/com/my/attendance/domain/Company.java
@@ -7,11 +7,11 @@ public class Company {
 	private String companyId;
 	private String comPw;
 	private String comName;
-	private int comRegno;
+	private long comRegno;
 	private String comAddr;
 	private String comDetailAddr;
 	private int comPostcode;
 	private String comPh;
 	private String comEmail;
-	private String president;
+	private String president;	
 }
